+++ lilypad/ee/server/models/__init__.py
@@ -1,14 +1,12 @@
 """The module for the `lilypad` EE database tables and models."""
 
-from .annotations import (
-    AnnotationBase,
-    AnnotationTable,
-)
+from .annotations import AnnotationBase, AnnotationTable, Label
 from .user_organizations import UserOrganizationBase, UserOrganizationTable, UserRole
 
 __all__ = [
     "AnnotationBase",
     "AnnotationTable",
+    "Label",
     "UserOrganizationBase",
     "UserOrganizationTable",
     "UserRole",
