+++ lilypad/ee/server/schemas/__init__.py
@@ -1,17 +1,1 @@
 """The module for the `lilypad` EE schemas."""
-
-from .annotations import AnnotationCreate, AnnotationPublic, AnnotationUpdate
-from .user_organizations import (
-    UserOrganizationCreate,
-    UserOrganizationPublic,
-    UserOrganizationUpdate,
-)
-
-__all__ = [
-    "AnnotationCreate",
-    "AnnotationPublic",
-    "AnnotationUpdate",
-    "UserOrganizationCreate",
-    "UserOrganizationPublic",
-    "UserOrganizationUpdate",
-]
