+++ lilypad/server/api/v0/api_keys_api.py
@@ -7,8 +7,8 @@ from uuid import UUID
 from fastapi import APIRouter, Depends
 
 from ..._utils import create_api_key
-from ...models import APIKeyTable
-from ...schemas import APIKeyCreate, APIKeyPublic
+from ...models.api_keys import APIKeyTable
+from ...schemas.api_keys import APIKeyCreate, APIKeyPublic
 from ...services import APIKeyService
 
 api_keys_api = APIRouter()
