+++ lilypad/server/api/v0/auth/github_api.py
@@ -9,7 +9,7 @@ from sqlmodel import Session
 
 from ...._utils.posthog import get_posthog_client
 from ....db import get_session
-from ....schemas import UserPublic
+from ....schemas.users import UserPublic
 from ....settings import Settings, get_settings
 from .utils import handle_user
