+++ lilypad/server/api/v0/spans_api.py
@@ -7,7 +7,8 @@ from uuid import UUID
 from fastapi import APIRouter, Depends, HTTPException
 
 from ...models import SpanTable
-from ...schemas import SpanMoreDetails, SpanPublic, SpanUpdate
+from ...schemas.span_more_details import SpanMoreDetails
+from ...schemas.spans import SpanPublic, SpanUpdate
 from ...services.spans import AggregateMetrics, SpanService, TimeFrame
 
 spans_router = APIRouter()
