+++ lilypad/server/services/base.py
@@ -11,7 +11,7 @@ from sqlmodel import Session, select
 from .._utils import get_current_user
 from ..db import get_session
 from ..models import BaseSQLModel
-from ..schemas import UserPublic
+from ..schemas.users import UserPublic
 
 _TableT = TypeVar("_TableT", bound=BaseSQLModel)
 _CreateT = TypeVar("_CreateT", bound=BaseModel)
