+++ lilypad/server/services/projects.py
@@ -1,7 +1,7 @@
 """The `ProjectService` class for projects."""
 
-from ..models import ProjectTable
-from ..schemas import ProjectCreate
+from ..models.projects import ProjectTable
+from ..schemas.projects import ProjectCreate
 from .base_organization import BaseOrganizationService
