+++ lilypad/server/services/tags.py
@@ -1,7 +1,7 @@
 """The `TagService` class for tags."""
 
-from ..models import TagTable
-from ..schemas import TagCreate
+from ..models.tags import TagTable
+from ..schemas.tags import TagCreate
 from .base_organization import BaseOrganizationService
