+++ lilypad/server/services/user_external_api_key_service.py
@@ -11,7 +11,7 @@ from .._utils.audit_logger import AuditAction, AuditLogger
 from ..db import get_session
 from ..models import UserTable
 from ..models.external_api_keys import ExternalAPIKeyTable
-from ..schemas import UserPublic
+from ..schemas.users import UserPublic
 from ..secret_manager.secret_manager_factory import get_secret_manager
 from ..services.base_organization import BaseOrganizationService
 
@@ -29,7 +29,6 @@ class UserExternalAPIKeyService(BaseOrganizationService):
         self.audit_logger = AuditLogger(session)
         # Initialize SecretManager instance (e.g., SupabaseVaultManager)
         self.secret_manager = get_secret_manager(self.session)
-
 
     def get_secret_name(self, service_name: str) -> str:
         """Generate a unique secret name for the user and service."""
@@ -83,19 +82,13 @@ class UserExternalAPIKeyService(BaseOrganizationService):
         name = self.get_secret_name(service_name)
         # Create a new secret in SecretManager and store its secret_id in the DB
         try:
-            secret_id = self.secret_manager.store_secret(
-                name, api_key, description
-            )
+            secret_id = self.secret_manager.store_secret(name, api_key, description)
         except IntegrityError:
             # If a duplicate key error occurs, delete the existing secret and retry storing
-            duplicate_secret_id = self.secret_manager.get_secret_id_by_name(
-                name
-            )
+            duplicate_secret_id = self.secret_manager.get_secret_id_by_name(name)
             if duplicate_secret_id:
                 self.secret_manager.delete_secret(duplicate_secret_id)
-            secret_id = self.secret_manager.store_secret(
-                name, api_key, description
-            )
+            secret_id = self.secret_manager.store_secret(name, api_key, description)
 
         new_key = ExternalAPIKeyTable(
             user_id=user.uuid,  # pyright: ignore [reportArgumentType]
