+++ tests/server/conftest.py
@@ -7,8 +7,10 @@ from sqlmodel import Session, SQLModel, create_engine
 
 from lilypad.ee.server.models import UserRole
 from lilypad.ee.server.schemas.user_organizations import UserOrganizationPublic
-from lilypad.server.schemas import (
+from lilypad.server.schemas.organizations import (
     OrganizationPublic,
+)
+from lilypad.server.schemas.users import (
     UserPublic,
 )
